refactor: reorganize and rename dragon-fight health management components
- Renamed `health-manager` to `dragon-fight/health-manager` for consistency.
- Moved `heal_dragon` logic to `dragon-fight/draconic-heart` and updated function naming.
- Updated import paths to reflect new structure.
