+++ behaviour_pack/scripts-dev/components/index.ts
@@ -3,7 +3,7 @@ import load_glitch_component from "./glitch";
 import load_altar_component from "./sacrifice";
 import load_whoop_component from "./whoop";
 import load_reactor_activate_component from "./reactor_activate";
-import load_heal_dragon_component from "./heal_dragon";
+import load_heal_dragon_component from "../features/dragon-fight/draconic-heart";
 import load_alcohol_component from "./alcohol";
 
 export default function load_custom_components(guild_id: string) {
