+++ behaviour_pack/scripts-dev/features/dragon-fight/index.ts
@@ -1,0 +1,7 @@
+import loadHealthManager from "./health-manager";
+import loadDraconicHeartComponents from "./draconic-heart";
+
+export default function loadDragonFightFeature() {
+    loadHealthManager()
+    loadDraconicHeartComponents()
+}
